fix: add dark mode, rename files
