+++ src/components/FilledButton.scss
@@ -48,11 +48,11 @@
         --back-color: var(--color-danger);
 
         &:hover {
-          --back-color: #d65550;
+          --back-color: var(--color-danger-darker);
         }
 
         &:active {
-          --back-color: #d1413c;
+          --back-color: var(--color-danger-darkest);
         }
       }
 
@@ -63,20 +63,20 @@
         --back-color: var(--input-bg-color);
 
         &:hover {
-          --text-color: #d65550;
-          --border-color: #d65550;
+          --text-color: var(--color-danger-darker);
+          --border-color: var(--color-danger-darker);
         }
 
         &:active {
-          --text-color: #d1413c;
-          --border-color: #d1413c;
+          --text-color: var(--color-danger-darkest);
+          --border-color: var(--color-danger-darkest);
         }
       }
     }
 
     &--color-muted {
       &.ExcButton--variant-filled {
-        --text-color: var(--input-bg-color);
+        --text-color: var(--island-bg-color);
         --back-color: var(--color-gray-50);
 
         &:hover {
@@ -90,18 +90,18 @@
 
       &.ExcButton--variant-outlined,
       &.ExcButton--variant-icon {
-        --text-color: var(--color-gray-80);
-        --border-color: var(--color-gray-30);
-        --back-color: var(--input-bg-color);
+        --text-color: var(--color-muted-background);
+        --border-color: var(--color-muted);
+        --back-color: var(--island-bg-color);
 
         &:hover {
-          --text-color: var(--color-gray-100);
-          --border-color: var(--color-gray-60);
+          --text-color: var(--color-muted-background-darker);
+          --border-color: var(--color-muted-darker);
         }
 
         &:active {
-          --text-color: var(--color-gray-100);
-          --border-color: var(--color-gray-100);
+          --text-color: var(--color-muted-background-darker);
+          --border-color: var(--color-muted-darkest);
         }
       }
     }
