+++ src/components/ShareableLinkDialog.scss
@@ -19,8 +19,6 @@
       line-height: 130%;
 
       margin: 0;
-
-      // TODO
     }
 
     &__popover {
