+++ src/css/theme.scss
@@ -102,6 +102,18 @@
   --color-text-warning: var(--text-primary-color);
 
   --color-danger: #db6965;
+  --color-danger-darker: #d65550;
+  --color-danger-darkest: #d1413c;
+  --color-danger-background: #fff0f0;
+  --color-danger-background-darker: #ffdad6;
+  --color-danger-text: #700000;
+
+  --color-muted: var(--color-gray-30);
+  --color-muted-darker: var(--color-gray-60);
+  --color-muted-darkest: var(--color-gray-100);
+  --color-muted-background: var(--color-gray-80);
+  --color-muted-background-darker: var(--color-gray-100);
+
   --color-promo: #e70078;
   --color-success: #268029;
   --color-success-lighter: #cafccc;
@@ -177,6 +189,18 @@
     --color-text-warning: var(--color-gray-80);
 
     --color-danger: #ffa8a5;
+    --color-danger-darker: #fd817c;
+    --color-danger-darkest: #e47571;
+    --color-danger-background: #261919;
+    --color-danger-background-darker: #452522;
+    --color-danger-text: #ffc4c4;
+
+    --color-muted: var(--color-gray-80);
+    --color-muted-darker: var(--color-gray-60);
+    --color-muted-darkest: var(--color-gray-20);
+    --color-muted-background: var(--color-gray-40);
+    --color-muted-background-darker: var(--color-gray-20);
+
     --color-promo: #d297ff;
   }
 }
